Concluído exercício 03 e ajustada definição de variáveis do exercício 02 do lab 03.

+++ Laboratorios/lab-03_2026-09-01/atividade-02/app/src/main/java/ads/poo/Horario.java
@@ -169,11 +169,11 @@ public class Horario {
         }
     }
 
-    public int converterParaSegundos() {
+    public long converterParaSegundos() {
         return (horas * 3600 + minutos * 60 + segundos);
     }
 
-    public int diferencaHorario(Horario horario) {
+    public long diferencaHorario(Horario horario) {
         if (this.converterParaSegundos() > horario.converterParaSegundos()) {
             return (86400 - this.converterParaSegundos() + horario.converterParaSegundos());
         } else {
